pyproject.toml
[project]
name = "app"
version = "0.2.0"
description = "Agentic RAG chatbot with PDF ingestion, grounded citations, and persistent memory"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "google-genai>=1.63.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
    "unstructured[all-docs]>=0.20.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that call external APIs (Gemini embeddings, etc.)",
]
